requirements.txt
asyncio==3.4.3
hikari==2.0.0.dev109
hikari-lightbulb==2.2.2
pyee==8.1.0
pyparsing==3.0.9
pyppeteer @ git+https://github.com/pyppeteer/pyppeteer@f98f9030d3d40dc629bf08f93672f05caf8dfb0e
pyppeteer-stealth==2.7.3
python-dotenv==0.20.0
python-twitter-v2==0.7.8
solana==0.25.0
virtualenv==20.16.2
vaderSentiment==3.3.2
